Added teacher dashboard and homework updaters

+++ backend/core/models.py
@@ -76,7 +76,7 @@ class Student(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE, related_name='student_profile')
     school = models.ForeignKey(School, related_name='students', on_delete=models.CASCADE)
     name = models.CharField(max_length=255)
-    effort_symbol = models.IntegerField(default=0)
+    effort_symbol = models.DecimalField(max_digits=5, decimal_places=2, default=0.00)
     words = models.ManyToManyField('Word', related_name='students', blank=True)
 
     def __str__(self):
